Possible SendImageInChat Fix
Sends a string instead of a byte array, with the starting number of each line labeled with a number. Then breaks apart the string into lines, takes a substring excluding the labeled number at the beginning, and converts the line back into a byte. Finally, makes the list of collected bytes into a byte array.